chore: 컨트롤러가 분리되어 테스트 코드도 분리
